Migration to the main.py file
